ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-anthropic
langchain-openai
langgraph
python-dotenv
pydantic
httpx
redis
pinecone
sentence-transformers
langchain-google-genai